Make *dataFile and interface
There was a lot of places that reached into a dataFile and accessed
private state.  This converts *dataFile to an interface to ensure
dataFile accesses go throught a well-defined interface.  This will
all make it easier to test the engine code that uses a dataFile
without having to actually create and write them on the file system.
